% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotmlx.R
\name{ggplotmlx}
\alias{ggplotmlx}
\title{mlxR wrapper for ggplot}
\usage{
ggplotmlx(...)
}
\arguments{
\item{\dots}{parameters passed to \code{\link[ggplot2]{ggplot}}}
}
\value{
see \code{\link[ggplot2]{ggplot}}
}
\description{
mlxR wrapper around \code{\link[ggplot2]{ggplot}} with a custom theme
}

