% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineDataFrame.R
\name{inlineDataFrame}
\alias{inlineDataFrame}
\title{inline data frame}
\usage{
inlineDataFrame(str, header = TRUE, colClasses = NA, ...)
}
\arguments{
\item{str}{text representation of the data frame}

\item{header}{see \code{\link{read.table}}}

\item{colClasses}{see \code{\link{read.table}}}

\item{\dots}{see \code{\link{read.table}}}
}
\description{
utility function to inline creation of a data frame
}

