% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{kmplotmlx}
\alias{kmplotmlx}
\title{brief description}
\usage{
kmplotmlx(r, index = 1, interval.censored = FALSE, level = NULL)
}
\arguments{
\item{r}{...}

\item{index}{...}

\item{interval.censored}{...}

\item{level}{}
}
\value{
what the function returns
}
\description{
longer description
possibly on multiple lines
}

