% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outputToList.R
\docType{data}
\name{list}
\alias{list}
\title{support for matlab-like multi results}
\format{\preformatted{Class 'result'  logi NA
}}
\usage{
list
}
\description{
support for matlab-like multi results
}
\keyword{datasets}

