% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxComputeRCleaner.R
\name{mlxComputeR.Unload}
\alias{mlxComputeR.Unload}
\title{mlxComputeRCleaner manages several sessions of mlxComputeR}
\usage{
mlxComputeR.Unload(arg = NULL)
}
\arguments{
\item{arg}{argument}
}
\description{
mlxComputeRCleaner manages several sessions of mlxComputeR
}

