% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxtoolboxReadData.R
\name{mlxtoolboxReadData}
\alias{mlxtoolboxReadData}
\title{Reads a data file in a monolix Project format for the simulator  Simulx using mlxtoobox code.}
\usage{
mlxtoolboxReadData(dataFile, colTypes)
}
\arguments{
\item{dataFile}{: the name (path) of the data file}

\item{colTypes}{: vector of strings, the types of different column present in the data file,
a column of a type not considered in mlxtoolbox or whose values is not desired can be declared as "IGNORE"}
}
\value{
datas: a list of containg:
\itemize{
  \item \code{sources} :   contains the treatment informations,
  \item \code{observation} : contains informations on observations,
  \item \code{covariates} : contains the covariates parameters.
}
}
\description{
Reads a data file in a monolix Project format for the simulator  Simulx using mlxtoobox code.
}
