% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monolix2simulx.R
\name{monolix2simulx}
\alias{monolix2simulx}
\title{Convert a Monolix Project  into an executable R script for the simulator  Simulx}
\usage{
monolix2simulx(project, parameter = NULL)
}
\arguments{
\item{project}{: the name of a Monolix project}

\item{parameter}{: string, the type of specific parameters to use: "mode" or "mean"...}
}
\value{
creates a folder containing files :
\itemize{
  \item \code{project.R}:  executable R code for the simulator,
  \item \code{treatment.txt}:   dose design informations,
  \item \code{parameters.txt}: population parameters ouput from Monolix,
  \item \code{group.txt}: group informations,
  \item \code{outputk.txt}: informations for the k-th output  (time, id),
  \item \code{covariates.txt}: individual covariates,
  \item \code{mode.txt} or \code{mean.txt}: specific individual parameters used for the simulation.
}
}
\description{
Convert a Monolix Project  into an executable R script for the simulator  Simulx
}

