% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmml2mlxtran.R
\name{pharmml2mlxtran}
\alias{pharmml2mlxtran}
\title{Translate PharmML into Mlxtran}
\usage{
pharmml2mlxtran(model.input, model.output = NULL)
}
\arguments{
\item{model.input}{the PharmML model (string)}

\item{model.output}{the Mlxtran model (string)}
}
\value{
the Mlxtran model (string)
}
\description{
Translate PharmML into Mlxtran
}

