% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.vector.R
\name{read.vector}
\alias{read.vector}
\title{Reads  a table into a vector}
\usage{
read.vector(f, header = FALSE, sep = "", quote = "\\"'")
}
\arguments{
\item{f}{: the table}

\item{header}{: bool, use the header or not}

\item{sep}{: the separator}

\item{quote}{: the quote character}
}
\value{
the vector
 @export
}
\description{
Reads  a table into a vector
}

