% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initMlxLibrary.R
\name{setMlxLibraryPath}
\alias{setMlxLibraryPath}
\title{Sets the MlxLibrary path in order to tell mlxR where is the MlxLibrary to use}
\usage{
setMlxLibraryPath(mlxLibraryPath)
}
\arguments{
\item{mlxLibraryPath}{the absolute path to the location of MlxLibrary}
}
\description{
Sets the MlxLibrary path in order to tell mlxR where is the MlxLibrary to use
}

