% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinymlx.R
\name{shinymlx}
\alias{shinymlx}
\title{Automatic code generation for Shiny applications}
\usage{
shinymlx(model, parameter = NULL, output = NULL, treatment = NULL,
  data = NULL)
}
\arguments{
\item{model}{a \code{Mlxtran} or \code{PharmML} model used for the simulation}

\item{parameter}{a vector of parameters with their names and values, or vectors (value, min, max, step)}

\item{output}{a list with fields:
\itemize{
  \item \code{name}: a vector of output names
  \item \code{time}: a vector of times, or a vector (min, max, step)
}}

\item{treatment}{a list with fields
\itemize{
  \item \code{tfd} : first time of dose (value, min, max, step),
  \item \code{amount} : amount (value, min, max, step),
  \item \code{nd} : number of doses (value, min, max, step),
  \item \code{ii} : interdose interval (value, min max, step),
  \item \code{type} : the type of input (value, max),
}
If no sliders need to be created for \code{treatment},
input argument of Simulx can also be used,  i.e. a list with fields
\itemize{
  \item \code{time} : a vector of input times,
  \item \code{amount} : a scalar or a vector of amounts,
  \item \code{rate} : a scalar or a vector of infusion rates (default=\code{Inf}),
  \item \code{tinf} : a scalar or a vector of infusion times (default=0),
  \item \code{type} : the type of input (default=1),
  \item \code{target} : the target compartment (default=NULL).
}}

\item{data}{a datafile to display with the plot}
}
\value{
A directory temp_dir with files ui.R, server.R and temp_model.txt
}
\description{
Creates a Shiny application for longitudinal data model
}
\details{
shinymlx automatically generates files ui.R and server.R required
for a Shiny application.

See http://simulx.webpopix.org for more details.
}

