% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmlx.R
\name{statmlx}
\alias{statmlx}
\title{Summary of data}
\usage{
statmlx(r, type = "continuous", FUN = "mean", probs = c(0.05, 0.5, 0.95),
  surv.time = NULL)
}
\arguments{
\item{r}{a data frame}

\item{type}{a string, type = {"continuous", "event"}. Default = "continuous"}

\item{probs}{a vector of quantiles  between 0 and 1. Only used when "quantile" has been defined in \code{FUN}}

\item{surv.time}{a scalar or a vector of times. Only used when "event" has been defined in \code{type}}

\item{stats}{a string, or a list of strings, with the name of the functions to apply to the result of the simulation}
}
\value{
A data frame.
}
\description{
Compute statistical summaries (mean, quantile, variance, survival rate,...)
}
\details{
See http://simulx.webpopix.org/stamlx for more details.
}

