% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDatamlx.R
\name{writeDatamlx}
\alias{writeDatamlx}
\title{Write formatted data file}
\usage{
writeDatamlx(r, result.file = NULL, result.folder = NULL, sep = ",",
  ext = NULL, digits = 5, app.file = F, app.dir = F)
}
\arguments{
\item{r}{a list of dataframes}

\item{result.file}{a string with the name of the file}

\item{result.folder}{a string with the name of the folder}

\item{sep}{(default = ",")}

\item{ext}{a string with the extension of the file names}

\item{digit}{(default = 5)}
}
\description{
Write data contained in a list of dataframes in a single file (NONMEM/Monolix format)
or in several files as tables
}
\details{
See http://simulx.webpopix.org/mlxr/writedatamlx/ for more details.
}
\examples{
\dontrun{

}
}

