% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{Computation of AUC, Cmax and Cmin}
\usage{
exposure(model, output, group = NULL, treatment = NULL, parameter = NULL,
  data = NULL, project = NULL, settings = NULL, regressor = NULL,
  varlevel = NULL)
}
\arguments{
\item{output}{a list  with fields:
\itemize{
  \item \code{name}: a vector of output names
  \item \code{time}: = 'steady.state' (mandatory)
  \item \code{ntp}: number of time points used for computing the exposure (default=100)
  \item \code{tol}: tolerance number, between 0 and 1, for approximating steaty-state  (default=0.01)
  \item \code{ngc}: number of doses used for estimating the convergence rate to steaty-state  (default=4)
}}

\item{treatment}{a list with fields
\itemize{
  \item \code{tfd} : time of first dose (default=0),
  \item \code{ii} : inter dose interval (mandatory),
  \item \code{amount} : the amount for each dose,
  \item \code{rate} : the infusion rate  (default=\code{Inf}),
  \item \code{tinf} : the time of infusion  (default=0),
  \item \code{type} : the type of input (default=1),
  \item \code{target} : the target compartment (default=NULL).
}}
}
\value{
A list of data frames
}
\description{
Compute the area under the curve, the maximum and minimum values of a fucntion of time over an interval
}
\details{
Input arguments are the input arguments of Simulx

Specific input arguments can be also used for computing the exposure at steady state,
i.e. after the administration of an "infinite" number of doses.  ()

See http://simulx.webpopix.org for more details.
}

