% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inlineModel.R
\name{inlineModel}
\alias{inlineModel}
\title{inline model}
\usage{
inlineModel(str, filename = NULL)
}
\arguments{
\item{str}{model}

\item{filename}{where to write the temporary model}
}
\description{
inline model
}

