% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monolix2simulx.R
\name{monolix2simulx}
\alias{monolix2simulx}
\title{Convert a Monolix Project  into an executable for the simulator  Simulx}
\usage{
monolix2simulx(project)
}
\arguments{
\item{projectName}{: the name of a Monolix project}
}
\value{
creates a folder projectNameR  containing files :
\itemize{
  \item \code{projectName.R} :  executable R code for the simulator,
  \item \code{treatment.txt} :  contains the treatment informations,
  \item \code{parameters.txt} : contains the  parameters ouput from Monolix,
  \item \code{group.txt} : contains the group informations
  \item \code{output.info} : contains the informations on the number of output of the mlxtran model
  \item \code{output_i.txt} : contains the output number i informations (time, id).
}

A list of data frames. Each data frame is an output of simulxthe mlxtran model
the data inputs: treatment, parameters, output of monolix, group...
}
\description{
Convert a Monolix Project  into an executable for the simulator  Simulx
}

