% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monolix2simulx.R
\name{monolix2simulx}
\alias{monolix2simulx}
\title{Convert a Monolix Project  into an executable for the simulator  Simulx}
\usage{
monolix2simulx(project, parameter = NULL)
}
\arguments{
\item{project}{: the name of a Monolix project}

\item{parameter}{: string $(NameOfTypeOfParameter), the type of specific parameters to use
example: "mode", "mean"...}
}
\value{
creates a folder projectNameR  containing files :
\itemize{
  \item \code{projectName.R} :  executable R code for the simulator,
  \item \code{treatment.txt} :  contains the treatment informations,
  \item \code{parameters.txt} : contains the  population parameters ouput from Monolix,
  \item \code{group.txt} : contains the group informations,
  \item \code{outputi.txt} : contains the output number i informations (time, id),
  \item \code{covariates.txt} : contains the covariates parameters,
  \item \code{$(NameOfTypeOfParameter)s.txt} : contains the specific parameter used.
}

A list of data frames. Each data frame is an output of simulx the mlxtran model
the data inputs: treatment, parameters, output of monolix, group...
}
\description{
Convert a Monolix Project  into an executable for the simulator  Simulx
}

