% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.vector.R
\name{read.vector}
\alias{read.vector}
\title{Reads  a table into a vector}
\usage{
read.vector(f, header = FALSE, sep = "", quote = "\\"'")
}
\arguments{
\item{f:}{the table}
}
\value{
the vector
}
\description{
Reads  a table into a vector
}

